Use command line options or inventory file when running the installer
Changes allow to use either command line options with the awx installer
or change the settings within the inventory file. The later seems to be
the better option, as one could remove lines like `dockerhub_base` and
`dockerhub_latest` to build your own images. Build requirements are only
installed if those to options are disabled.

The changes involve a more sophisticated datastructure in the
defaults file.

This is a proof of concept and should be discussed in detail.